mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.14",
    "fastmcp>=2.10.1",
    "google-auth-oauthlib>=1.2.2",
    "pydantic-ai[mcp]>=0.3.5",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.2",
    "uvicorn>=0.35.0",
]
